fix(levelup):refactor IPitchRestInterface
Creating collection package, moving IPitchRestInterface into the Message
class and deleting IPitchRestinterface
